+++ src/pymmcore_plus/mda/metadata/__init__.py
@@ -1,0 +1,3 @@
+from ._base import MetadataProvider, ensure_valid_metadata_func, get_metadata_func
+
+__all__ = ["MetadataProvider", "get_metadata_func", "ensure_valid_metadata_func"]
