+++ src/pymmcore_plus/mda/metadata/_legacy.py
@@ -1,0 +1,4 @@
+from ._base import MetadataProvider
+
+
+class LegacySummaryProvider(MetadataProvider): ...
